# doc/contracts/node-protocol.openapi.yaml
openapi: 3.1.0
info:
  title: lingfeng-workbench Node Protocol
  version: 1.0.0
  description: HTTPS pull protocol independently implemented by Service and Node.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/node/v1
security:
  - bearerAuth: []
paths:
  /hello:
    post:
      operationId: registerNode
      summary: Register or refresh the authenticated Node projection
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/HelloRequest' } } }
      responses:
        '200': { description: Registered, content: { application/json: { schema: { $ref: '#/components/schemas/Acknowledgement' } } } }
        '400': { $ref: '#/components/responses/Rejected' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /heartbeat:
    post:
      operationId: heartbeatNode
      summary: Refresh liveness for the authenticated Node
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/HeartbeatRequest' } } }
      responses:
        '200': { description: Heartbeat accepted, content: { application/json: { schema: { $ref: '#/components/schemas/Acknowledgement' } } } }
        '400': { $ref: '#/components/responses/Rejected' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /poll:
    post:
      operationId: pollNodeCommand
      summary: Pull at most one durable assignment for the authenticated Node
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PollRequest' } } }
      responses:
        '200':
          description: At most one durable command
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/NoCommand'
                  - $ref: '#/components/schemas/AssignmentCommand'
        '400': { $ref: '#/components/responses/Rejected' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /events:
    post:
      operationId: recordNodeEvent
      summary: Record one idempotent normalized Run event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/RunAcceptedEvent'
                - $ref: '#/components/schemas/RunStartedEvent'
                - $ref: '#/components/schemas/ProgressEvent'
                - $ref: '#/components/schemas/TerminalEvent'
      responses:
        '200': { description: Event accepted or idempotently replayed, content: { application/json: { schema: { $ref: '#/components/schemas/Acknowledgement' } } } }
        '400': { $ref: '#/components/responses/Rejected' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  responses:
    Rejected:
      description: Malformed or unsupported message
      content: { application/json: { schema: { $ref: '#/components/schemas/ProtocolError' } } }
    Unauthorized:
      description: Invalid node credential
      content: { application/json: { schema: { $ref: '#/components/schemas/ProtocolError' } } }
    Forbidden:
      description: Credential is not bound to the claimed node
      content: { application/json: { schema: { $ref: '#/components/schemas/ProtocolError' } } }
    Conflict:
      description: Message conflicts with durable state
      content: { application/json: { schema: { $ref: '#/components/schemas/ProtocolError' } } }
  schemas:
    Identifier:
      type: string
      pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$'
    ShortText:
      type: string
      minLength: 1
      maxLength: 800
    Timestamp:
      type: string
      format: date-time
    MessageHeader:
      type: object
      required: [protocolVersion, messageId, nodeId, sentAt]
      properties:
        protocolVersion: { type: string, const: '1.0' }
        messageId: { $ref: '#/components/schemas/Identifier' }
        nodeId: { $ref: '#/components/schemas/Identifier' }
        sentAt: { $ref: '#/components/schemas/Timestamp' }
    HelloRequest:
      allOf:
        - $ref: '#/components/schemas/MessageHeader'
        - type: object
          required: [displayName, capabilities]
          properties:
            displayName: { $ref: '#/components/schemas/ShortText' }
            capabilities:
              type: array
              maxItems: 32
              items: { $ref: '#/components/schemas/Identifier' }
      unevaluatedProperties: false
    HeartbeatRequest:
      allOf:
        - $ref: '#/components/schemas/MessageHeader'
        - type: object
          properties:
            activeRunId: { $ref: '#/components/schemas/Identifier' }
      unevaluatedProperties: false
    PollRequest:
      allOf:
        - $ref: '#/components/schemas/MessageHeader'
        - type: object
          properties:
            acknowledgedCommandIds:
              type: array
              maxItems: 100
              items: { $ref: '#/components/schemas/Identifier' }
      unevaluatedProperties: false
    NoCommand:
      type: object
      additionalProperties: false
      required: [commandType]
      properties:
        commandType: { type: string, const: NO_COMMAND }
    AssignmentCommand:
      type: object
      additionalProperties: false
      required: [commandType, commandId, workItemId, missionId, runId, missionRevision, missionDigest, objective, acceptanceSummary, authorizedSideEffectsSummary, targetNodeId, workspaceRef, runtimeKind, executionProfile]
      properties:
        commandType: { type: string, const: ASSIGNMENT }
        commandId: { $ref: '#/components/schemas/Identifier' }
        workItemId: { $ref: '#/components/schemas/Identifier' }
        missionId: { $ref: '#/components/schemas/Identifier' }
        runId: { $ref: '#/components/schemas/Identifier' }
        missionRevision: { type: integer, minimum: 1 }
        missionDigest: { type: string, pattern: '^[a-f0-9]{64}$' }
        objective: { $ref: '#/components/schemas/ShortText' }
        acceptanceSummary: { $ref: '#/components/schemas/ShortText' }
        authorizedSideEffectsSummary: { $ref: '#/components/schemas/ShortText' }
        targetNodeId: { $ref: '#/components/schemas/Identifier' }
        workspaceRef: { $ref: '#/components/schemas/Identifier' }
        runtimeKind: { $ref: '#/components/schemas/Identifier' }
        executionProfile: { $ref: '#/components/schemas/Identifier' }
    RunEventHeader:
      allOf:
        - $ref: '#/components/schemas/MessageHeader'
        - type: object
          required: [eventType, workItemId, missionId, runId, missionDigest]
          properties:
            eventType: { type: string }
            workItemId: { $ref: '#/components/schemas/Identifier' }
            missionId: { $ref: '#/components/schemas/Identifier' }
            runId: { $ref: '#/components/schemas/Identifier' }
            missionDigest: { type: string, pattern: '^[a-f0-9]{64}$' }
    RunAcceptedEvent:
      allOf:
        - $ref: '#/components/schemas/RunEventHeader'
        - type: object
          required: [eventType, commandId]
          properties:
            eventType: { type: string, const: RUN_ACCEPTED }
            commandId: { $ref: '#/components/schemas/Identifier' }
      unevaluatedProperties: false
    RunStartedEvent:
      allOf:
        - $ref: '#/components/schemas/RunEventHeader'
        - type: object
          required: [eventType, resumable]
          properties:
            eventType: { type: string, const: RUN_STARTED }
            resumable: { type: boolean }
      unevaluatedProperties: false
    ProgressEvent:
      allOf:
        - $ref: '#/components/schemas/RunEventHeader'
        - type: object
          required: [eventType, progressSummary]
          properties:
            eventType: { type: string, const: PROGRESS }
            progressSummary: { $ref: '#/components/schemas/ShortText' }
      unevaluatedProperties: false
    TerminalEvent:
      allOf:
        - $ref: '#/components/schemas/RunEventHeader'
        - type: object
          required: [eventType, runtimeOutcome, acceptanceStatus, resultSummary]
          properties:
            eventType: { type: string, enum: [EXECUTION_FINISHED, EXECUTION_FAILED, EXECUTION_INTERRUPTED] }
            runtimeOutcome: { type: string, enum: [SUCCEEDED, FAILED, INTERRUPTED] }
            acceptanceStatus: { type: string, enum: [PASSED, FAILED, UNKNOWN] }
            resultSummary: { $ref: '#/components/schemas/ShortText' }
      unevaluatedProperties: false
    Acknowledgement:
      type: object
      additionalProperties: false
      required: [requestMessageId, duplicate]
      properties:
        requestMessageId: { $ref: '#/components/schemas/Identifier' }
        duplicate: { type: boolean }
    ProtocolError:
      type: object
      additionalProperties: false
      required: [code, message, requestMessageId]
      properties:
        code: { $ref: '#/components/schemas/Identifier' }
        message: { $ref: '#/components/schemas/ShortText' }
        requestMessageId: { $ref: '#/components/schemas/Identifier' }